Move the status dot to the left of the date, spell out the weekday
fmtDate() now returns the full weekday name instead of the 3-letter
abbreviation, and the day-of-week line matches the month/day line's size
and font (28px, heading serif) instead of being smaller. The glowing
status dot moves from above the date block to its left, in a row layout.
Widened the date column (130px -> 190px) to fit "SATURDAY" comfortably —
verified by measuring actual rendered text width against the container
before pushing.

+++ app/board/ui.tsx
@@ -34,7 +34,7 @@ export function fmtDate(iso: string | null) {
   return {
     day: String(d.getDate()).padStart(2, "0"),
     mon: d.toLocaleString("en-US", { month: "short" }).toUpperCase(),
-    dow: d.toLocaleString("en-US", { weekday: "short" }),
+    dow: d.toLocaleString("en-US", { weekday: "long" }),
     year: d.getFullYear(),
   };
 }
